+++ src/Phan/AST/TolerantASTConverter/NodeDumper.php
@@ -145,7 +145,7 @@ class NodeDumper
                 \Phan\Library\StringUtil::jsonEncode(\substr($this->file_contents, $ast_node->fullStart, $ast_node->length))
             );
         } elseif (\is_scalar($ast_node)) {
-            return \var_export($ast_node, true);
+            return \var_representation($ast_node);
         } elseif ($ast_node === null) {
             return 'null';
         } else {