+++ tests/files/src/0936_null_narrowed_form_of_nullable.php
@@ -1,0 +1,19 @@
+<?php
+
+class Demo1 {
+    /**
+     * @return ?int
+     */
+    public function func() : ?int {
+        return 1;
+    }
+}
+
+class Demo2 extends Demo1 {
+    /**
+     * @return null
+     */
+    public function func() : ?int {
+        return null;
+    }
+}